backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
opencv-python>=4.8.0
ultralytics>=8.1.0
numpy>=1.24.0
werkzeug>=3.0.0
matplotlib>=3.8.0
pandas>=2.2.0
scikit-learn>=1.4.0
python-dotenv>=1.0.1
pymongo>=4.6.0
requests>=2.31.0

